# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MTB REST API
  version: 1.0.0
  description: REST API of the MTB service.
...
servers:
  - url: /api
paths:
...
  /upload:
    post:
      summary: Receive and save files on the file system
      operationId: uploadFile
      tags:
        - upload
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                # 'file' will be the field name in this multipart request
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        405:
          description: Invalid Input
...